+++ geotrek/maintenance/tests/test_filters.py
@@ -1,7 +1,7 @@
 from unittest import skipIf
 
 from django.conf import settings
-from django.contrib.gis.geos import LineString, MultiPolygon, Polygon
+from django.contrib.gis.geos import GeometryCollection, LineString, MultiPolygon, Point, Polygon
 from django.test import TestCase
 
 from geotrek.land.tests.factories import (
@@ -15,11 +15,10 @@ from geotrek.land import filters  # noqa
 
 from geotrek.maintenance.filters import ProjectFilterSet, InterventionFilterSet
 from geotrek.maintenance.tests.factories import InterventionFactory, ProjectFactory
+from geotrek.outdoor.tests.factories import SiteFactory, CourseFactory
+from geotrek.signage.tests.factories import BladeFactory, SignageFactory
 from geotrek.zoning.tests.factories import (CityFactory, DistrictFactory,
                                             RestrictedAreaFactory, RestrictedAreaTypeFactory)
-
-if 'geotrek.outdoor' in settings.INSTALLED_APPS:
-    from geotrek.outdoor.tests.factories import SiteFactory, CourseFactory
 
 
 class InterventionFilteringByBboxTest(TestCase):
@@ -271,6 +270,288 @@ class ProjectFilteringByLandTest(TestCase):
 
         self.assertEqual(len(qs), 1)
         self.assertEqual(qs[0], self.seek_proj)
+
+
+class InterventionIntersectionFilterZoningTest(TestCase):
+    @classmethod
+    def setUpTestData(cls):
+        if settings.TREKKING_TOPOLOGY_ENABLED:
+            cls.path_in_1 = PathFactory.create(geom=LineString((0, 0), (2, 1), srid=settings.SRID))
+            cls.path_in_2 = PathFactory.create(geom=LineString((5, 5), (4, 4), srid=settings.SRID))
+            cls.topo_in_1 = TopologyFactory.create(paths=[cls.path_in_1])
+            cls.topo_in_2 = TopologyFactory.create(paths=[cls.path_in_2])
+            signage_in_1 = SignageFactory.create(paths=[cls.path_in_1])
+            signage_in_2 = SignageFactory.create(paths=[cls.path_in_2])
+        else:
+            cls.topo_in_1 = TopologyFactory.create(geom=LineString((0, 0), (2, 1), srid=settings.SRID))
+            cls.topo_in_2 = TopologyFactory.create(geom=LineString((5, 5), (4, 4), srid=settings.SRID))
+            signage_in_1 = SignageFactory.create(geom=Point(1, 1, srid=settings.SRID))
+            signage_in_2 = SignageFactory.create(geom=Point(5, 5, srid=settings.SRID))
+        cls.intervention_topology_in_1 = InterventionFactory.create(target=cls.topo_in_1)
+        cls.intervention_topology_in_2 = InterventionFactory.create(target=cls.topo_in_2)
+
+        cls.site_in_1 = SiteFactory.create(geom=GeometryCollection(Point(1, 1), srid=settings.SRID))
+        cls.intervention_site_in_1 = InterventionFactory.create(target=cls.site_in_1)
+
+        cls.course_in_1 = CourseFactory.create(parent_sites=[cls.site_in_1.pk],
+                                               geom=GeometryCollection(Point(1, 1), srid=settings.SRID))
+        cls.intervention_course_in_1 = InterventionFactory.create(target=cls.course_in_1)
+
+        cls.site_in_2 = SiteFactory.create(geom=GeometryCollection(Point(5, 5), srid=settings.SRID))
+        cls.intervention_site_in_2 = InterventionFactory.create(target=cls.site_in_2)
+
+        cls.course_in_2 = CourseFactory.create(parent_sites=[cls.site_in_2.pk],
+                                               geom=GeometryCollection(Point(5, 5), srid=settings.SRID))
+        cls.intervention_course_in_2 = InterventionFactory.create(target=cls.course_in_2)
+
+        cls.blade_in_1 = BladeFactory.create(signage=signage_in_1)
+        cls.intervention_blade_in_1 = InterventionFactory.create(target=cls.blade_in_1)
+
+        cls.blade_in_2 = BladeFactory.create(signage=signage_in_2)
+        cls.intervention_blade_in_2 = InterventionFactory.create(target=cls.blade_in_2)
+
+        cls.geom_in_1 = MultiPolygon(Polygon(((0, 0), (2, 0), (2, 2), (0, 2), (0, 0)), srid=settings.SRID))
+        cls.geom_in_2 = MultiPolygon(Polygon(((4, 4), (5, 4), (5, 5), (4, 5), (4, 4)), srid=settings.SRID))
+        cls.geom_out = MultiPolygon(Polygon(((6, 6), (10, 6), (10, 10), (6, 10), (6, 6)), srid=settings.SRID))
+
+    def test_filter_in_1_city(self):
+        """
+        We should have 1 interventions on topologies, 1 intervention on sites, 1 intervention on courses,
+        1 intervention on blade
+        """
+        filter = InterventionFilterSet(data={'city': [CityFactory.create(geom=self.geom_in_1)]})
+        self.assertTrue(filter.is_valid())
+        self.assertSetEqual(set(filter.qs),
+                            {self.intervention_topology_in_1,
+                             self.intervention_site_in_1,
+                             self.intervention_course_in_1,
+                             self.intervention_blade_in_1})
+        self.assertEqual(len(filter.qs), 4)
+
+    def test_filter_in_2_city(self):
+        """
+        We should have 1 interventions on topologies, 1 intervention on sites, 1 intervention on courses,
+        1 intervention on blade
+        """
+        filter = InterventionFilterSet(data={'city': [CityFactory.create(geom=self.geom_in_2)]})
+        self.assertTrue(filter.is_valid())
+        self.assertSetEqual(set(filter.qs),
+                            {self.intervention_topology_in_2,
+                             self.intervention_site_in_2,
+                             self.intervention_course_in_2,
+                             self.intervention_blade_in_2})
+        self.assertEqual(len(filter.qs), 4)
+
+    def test_filter_in_1_and_2_city(self):
+        """
+        We should have 2 interventions on topologies, 2 interventions on sites, 2 interventions on courses,
+        2 interventions on blade
+        """
+        filter = InterventionFilterSet(data={'city': [CityFactory.create(geom=self.geom_in_2),
+                                                      CityFactory.create(geom=self.geom_in_1)]})
+        self.assertTrue(filter.is_valid())
+        self.assertSetEqual(set(filter.qs),
+                            {self.intervention_topology_in_2,
+                             self.intervention_site_in_2,
+                             self.intervention_course_in_2,
+                             self.intervention_blade_in_2,
+                             self.intervention_topology_in_1,
+                             self.intervention_site_in_1,
+                             self.intervention_course_in_1,
+                             self.intervention_blade_in_1})
+        self.assertEqual(len(filter.qs), 8)
+
+    def test_filter_out_city(self):
+        """
+        We should not have any interventions
+        """
+        filter = InterventionFilterSet(data={'city': [CityFactory.create(geom=self.geom_out)]})
+        self.assertTrue(filter.is_valid())
+        self.assertEqual(len(filter.qs), 0)
+
+    def test_filter_in_1_district(self):
+        """
+        We should have 1 interventions on topologies, 1 intervention on sites, 1 intervention on courses,
+        1 intervention on blade
+        """
+        filter = InterventionFilterSet(data={'district': [DistrictFactory.create(geom=self.geom_in_1)]})
+        self.assertTrue(filter.is_valid())
+        self.assertSetEqual(set(filter.qs),
+                            {self.intervention_topology_in_1,
+                             self.intervention_site_in_1,
+                             self.intervention_course_in_1,
+                             self.intervention_blade_in_1})
+        self.assertEqual(len(filter.qs), 4)
+
+    def test_filter_in_2_district(self):
+        """
+        We should have 1 interventions on topologies, 1 intervention on sites, 1 intervention on courses,
+        1 intervention on blade
+        """
+        filter = InterventionFilterSet(data={'district': [DistrictFactory.create(geom=self.geom_in_2)]})
+        self.assertTrue(filter.is_valid())
+        self.assertSetEqual(set(filter.qs),
+                            {self.intervention_topology_in_2,
+                             self.intervention_site_in_2,
+                             self.intervention_course_in_2,
+                             self.intervention_blade_in_2})
+        self.assertEqual(len(filter.qs), 4)
+
+    def test_filter_in_1_and_2_district(self):
+        """
+        We should have 2 interventions on topologies, 2 interventions on sites, 2 interventions on courses,
+        2 interventions on blade
+        """
+        filter = InterventionFilterSet(data={'district': [DistrictFactory.create(geom=self.geom_in_1),
+                                                          DistrictFactory.create(geom=self.geom_in_2)]})
+        self.assertTrue(filter.is_valid())
+        self.assertSetEqual(set(filter.qs),
+                            {self.intervention_topology_in_2,
+                             self.intervention_site_in_2,
+                             self.intervention_course_in_2,
+                             self.intervention_blade_in_2,
+                             self.intervention_topology_in_1,
+                             self.intervention_site_in_1,
+                             self.intervention_course_in_1,
+                             self.intervention_blade_in_1})
+        self.assertEqual(len(filter.qs), 8)
+
+    def test_filter_out_district(self):
+        """
+        We should not have any interventions
+        """
+        filter = InterventionFilterSet(data={'district': [DistrictFactory.create(geom=self.geom_out)]})
+        self.assertTrue(filter.is_valid())
+        self.assertEqual(len(filter.qs), 0)
+
+    def test_filter_in_1_restricted_area(self):
+        """
+        We should have 1 interventions on topologies, 1 intervention on sites, 1 intervention on courses,
+        1 intervention on blade
+        """
+        filter = InterventionFilterSet(data={'area': [RestrictedAreaFactory.create(geom=self.geom_in_1)]})
+        self.assertSetEqual(set(filter.qs),
+                            {self.intervention_topology_in_1,
+                             self.intervention_site_in_1,
+                             self.intervention_course_in_1,
+                             self.intervention_blade_in_1})
+        self.assertEqual(len(filter.qs), 4)
+
+    def test_filter_in_2_restricted_area(self):
+        """
+        We should have 1 interventions on topologies, 1 intervention on sites, 1 intervention on courses,
+        1 intervention on blade
+        """
+        filter = InterventionFilterSet(data={'area': [RestrictedAreaFactory.create(geom=self.geom_in_2)]})
+        self.assertTrue(filter.is_valid())
+        self.assertSetEqual(set(filter.qs),
+                            {self.intervention_topology_in_2,
+                             self.intervention_site_in_2,
+                             self.intervention_course_in_2,
+                             self.intervention_blade_in_2})
+        self.assertEqual(len(filter.qs), 4)
+
+    def test_filter_in_1_and_2_restricted_area(self):
+        """
+        We should have 2 interventions on topologies, 2 interventions on sites, 2 interventions on courses,
+        2 interventions on blade
+        """
+        filter = InterventionFilterSet(data={'area': [RestrictedAreaFactory.create(geom=self.geom_in_1),
+                                                      RestrictedAreaFactory.create(geom=self.geom_in_2)]})
+        self.assertTrue(filter.is_valid())
+        self.assertSetEqual(set(filter.qs),
+                            {self.intervention_topology_in_2,
+                             self.intervention_site_in_2,
+                             self.intervention_course_in_2,
+                             self.intervention_blade_in_2,
+                             self.intervention_topology_in_1,
+                             self.intervention_site_in_1,
+                             self.intervention_course_in_1,
+                             self.intervention_blade_in_1})
+        self.assertEqual(len(filter.qs), 8)
+
+    def test_filter_out_restricted_area(self):
+        """
+        We should not have any interventions
+        """
+        filter = InterventionFilterSet(data={'area': [RestrictedAreaFactory.create(geom=self.geom_out)]})
+        self.assertTrue(filter.is_valid())
+        self.assertEqual(len(filter.qs), 0)
+
+    def test_filter_in_1_restricted_area_type(self):
+        """
+        We should have 1 interventions on topologies, 1 intervention on sites, 1 intervention on courses,
+        1 intervention on blade
+        """
+        restricted_area_type = RestrictedAreaTypeFactory.create()
+        RestrictedAreaFactory.create(geom=self.geom_in_1, area_type=restricted_area_type)
+        filter = InterventionFilterSet(data={'area_type': [restricted_area_type.pk]})
+        self.assertTrue(filter.is_valid())
+        self.assertSetEqual(set(filter.qs),
+                            {self.intervention_topology_in_1,
+                             self.intervention_site_in_1,
+                             self.intervention_course_in_1,
+                             self.intervention_blade_in_1})
+        self.assertEqual(len(filter.qs), 4)
+
+    def test_filter_in_2_restricted_area_type(self):
+        """
+        We should have 1 interventions on topologies, 1 intervention on sites, 1 intervention on courses,
+        1 intervention on blade
+        """
+        restricted_area_type = RestrictedAreaTypeFactory.create()
+        RestrictedAreaFactory.create(geom=self.geom_in_2, area_type=restricted_area_type)
+        filter = InterventionFilterSet(data={'area_type': [restricted_area_type.pk]})
+        self.assertTrue(filter.is_valid())
+        self.assertSetEqual(set(filter.qs),
+                            {self.intervention_topology_in_2,
+                             self.intervention_site_in_2,
+                             self.intervention_course_in_2,
+                             self.intervention_blade_in_2})
+        self.assertEqual(len(filter.qs), 4)
+
+    def test_filter_in_1_and_in_2_restricted_area_type(self):
+        """
+        We should have 2 interventions on topologies, 2 interventions on sites, 2 interventions on courses,
+        2 interventions on blade
+        """
+        restricted_area_type_1 = RestrictedAreaTypeFactory.create()
+        restricted_area_type_2 = RestrictedAreaTypeFactory.create()
+
+        RestrictedAreaFactory.create(geom=self.geom_in_1, area_type=restricted_area_type_1)
+        RestrictedAreaFactory.create(geom=self.geom_in_2, area_type=restricted_area_type_2)
+        filter = InterventionFilterSet(data={'area_type': [restricted_area_type_1.pk,
+                                                           restricted_area_type_2.pk]})
+        self.assertTrue(filter.is_valid())
+        self.assertSetEqual(set(filter.qs),
+                            {self.intervention_topology_in_2,
+                             self.intervention_site_in_2,
+                             self.intervention_course_in_2,
+                             self.intervention_blade_in_2,
+                             self.intervention_topology_in_1,
+                             self.intervention_site_in_1,
+                             self.intervention_course_in_1,
+                             self.intervention_blade_in_1})
+        self.assertEqual(len(filter.qs), 8)
+
+    def test_filter_out_restricted_area_type(self):
+        """
+        We should not have any interventions
+        """
+        restricted_area_type = RestrictedAreaTypeFactory.create()
+        RestrictedAreaFactory.create(geom=self.geom_out, area_type=restricted_area_type)
+        filter = InterventionFilterSet(data={'area_type': [restricted_area_type.pk]})
+        self.assertTrue(filter.is_valid())
+        self.assertEqual(len(filter.qs), 0)
+
+    def test_filter_restricted_area_type_without_restricted_area(self):
+        """
+        We should not have any interventions
+        """
+        restricted_area_type = RestrictedAreaTypeFactory.create()
+        filter = ProjectFilterSet(data={'area_type': [restricted_area_type.pk]})
+        self.assertTrue(filter.is_valid())
+        self.assertEqual(len(filter.qs), 0)
 
 
 class ProjectIntersectionFilterZoningTest(TestCase):